+++ CRM/Dataprocessor/Form/Output/AbstractUIOutputForm.php
@@ -162,12 +162,36 @@ abstract class CRM_Dataprocessor_Form_Output_AbstractUIOutputForm extends CRM_Co
         if ($filter->isExposed()) {
           $filterValues = $filter->processSubmittedValues($submittedValues);
           if (empty($filterValues)) {
-            $filterValues = $filter->getDefaultFilterValues();
+            $filterValues = self::getDefaultFilterValues($filter);
           }
           $filter->applyFilterFromSubmittedFilterParams($filterValues);
         }
       }
     }
+  }
+
+  /**
+   * Get the default filter values for a filter. If there is no default value we allow the value to be set by a URL parameter of the same name as the filter.
+   *
+   * @param \Civi\DataProcessor\FilterHandler\AbstractFilterHandler $filterHandler
+   *
+   * @return array
+   * @throws \CRM_Core_Exception
+   */
+  public static function getDefaultFilterValues(\Civi\DataProcessor\FilterHandler\AbstractFilterHandler $filterHandler) {
+    $filterValues = $filterHandler->getDefaultFilterValues();
+    if (empty($filterValues)) {
+      $type = ($filterHandler->getFieldSpecification()->type === 'Int') ? 'CommaSeparatedIntegers' : $filterHandler->getFieldSpecification()->type;
+
+      $valueFromURL = \CRM_Utils_Request::retrieveValue($filterHandler->getFieldSpecification()->alias, $type, NULL, FALSE, 'GET');
+      if ($valueFromURL) {
+        $filterValues = [
+          'op' => 'IN',
+          'value' => $valueFromURL,
+        ];
+      }
+    }
+    return $filterValues;
   }
 
   /**
@@ -181,7 +205,7 @@ abstract class CRM_Dataprocessor_Form_Output_AbstractUIOutputForm extends CRM_Co
         if (!$fieldSpec || !$filterHandler->isExposed()) {
           continue;
         }
-        $filterElements[$fieldSpec->alias]['filter'] = $filterHandler->addToFilterForm($this, $filterHandler->getDefaultFilterValues(), $this->getCriteriaElementSize());
+        $filterElements[$fieldSpec->alias]['filter'] = $filterHandler->addToFilterForm($this, self::getDefaultFilterValues($filterHandler), $this->getCriteriaElementSize());
         $filterElements[$fieldSpec->alias]['template'] = $filterHandler->getTemplateFileName();
       }
       $this->assign('filters', $filterElements);